backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv==1.0.1
motor==3.3.1
pymongo==4.5.0
httpx==0.28.1
pydantic==2.6.4
email-validator==2.1.0.post1
google-auth==2.29.0
requests==2.32.3